backend/requirements.txt
Django==3.2
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.0.0
drf-extra-fields==3.7.0
django-filter==23.2
psycopg2-binary==2.9.6
gunicorn==21.2.0